+++ app-web/src/main/java/ar/com/educacionit/web/servlet/NuevoProductoServlet.java
@@ -5,7 +5,6 @@ import java.util.Collection;
 
 import javax.servlet.ServletException;
 import javax.servlet.annotation.WebServlet;
-import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 
@@ -32,7 +31,7 @@ public class NuevoProductoServlet extends BaseServlet {
 		
 		try {
 			
-			Float precioF = Float.parseFloat(precio);
+			Double precioF = Double.parseDouble(precio);
 			Long tipoProductoL = Long.parseLong(tipoProducto);
 			
 			Producto nuevoProducto = new Producto(titulo, precioF, codigo, tipoProductoL);
@@ -45,9 +44,9 @@ public class NuevoProductoServlet extends BaseServlet {
 			//recargo la lista de productos
 			Collection<Producto> productos = ps.findAll();
 			request.setAttribute(ViewKeyEnums.LISTADO.name(), productos);
-			Float suma = productos.stream()
+			Double suma = productos.stream()
 					.map(p -> p.getPrecio())
-					.reduce(0F, (Float x, Float y) -> x+y);
+					.reduce(0D, (Double x, Double y) -> x+y);
 				
 			addAttribute(request, ViewKeyEnums.TOTAL, suma);
